cambios Logo + responsi-definitivo
