+++ CLIENT/src/components/Anuncios/Anuncio2.jsx
@@ -10,49 +10,41 @@ import {
 } from "@chakra-ui/react";
 import React from "react";
 import imagelanding from "../../assets/image/imagen_landing1.jpg";
-import Logo2 from "../../assets/image/Logo2.png";
+import Logo2 from "../../assets/image/Logo2.1.png";
 import "animate.css";
 import { Link } from "react-router-dom";
 import { FaChevronDown } from "react-icons/fa";
-import "./Anuncio2.css"
+import "./Anuncio2.css";
+import "../../assets/fonts/Abril_Fatface/AbrilFatface-Regular.ttf";
 
 export default function Anuncio2() {
   const bg = {
     bg: "gray.100",
   };
-  const handleScroll = () => {
-    const currentPosition = window.pageYOffset;
-    window.scrollTo({
-      top: currentPosition + 750,
-      behavior: "smooth",
-    });
-  };
+ 
 
   return (
     <chakra.header>
       <Box
         w={["100%", "100%", "100%"]}
-        h= {["50vh", "75vh", "100vh"]}
+        h={["50vh", "75vh", "100vh"]}
         backgroundImage={imagelanding}
         bgPos="center"
         bgSize="cover"
       >
         <Flex
+          direction={["column", "column", "row"]}
           justify="center"
           pos="relative"
           boxSize="full"
           bg="blackAlpha.700"
         >
           <Box
-            w={["20%", "20%", "25%"]}
-            h={["20%", "20%", "25%"]}
-            border="1px solid black"
-            borderRadius="10px"
-            backgroundColor="#70a57b"
-            boxShadow={
-              "rgba(0, 0, 0, 0.4) 0px 2px 4px, rgba(0, 0, 0, 0.3) 0px 7px 13px -3px, rgba(0, 0, 0, 0.2) 0px -3px 0px inset"
-            }
+            w={["80%", "80%", "35%"]}
+            h={["80%", "80%", "35%"]}
             m={["8px", "8px", "20px"]}
+    
+
             display="flex"
             justifyContent="center"
             alignItems="center"
@@ -70,11 +62,16 @@ export default function Anuncio2() {
           >
             <Stack spacing={20} textAlign="center">
               <Heading
-                fontSize={["2xl", "3xl", "6xl"]}
+                fontSize={["4xl", "4xl", "6xl"]}
+                fontFamily="Abril Fatface"
                 fontWeight="semibold"
                 color="brand.primary"
                 textTransform="uppercase"
                 className="animate__animated animate__zoomInDown animate__delay-1s"
+                style={{
+                  textShadow: "4px 4px 16px rgba(0, 0, 0, 0.3)",
+                  textOutline: "3px solid white",
+                }}
               >
                 Bienvenidos
               </Heading>
@@ -83,7 +80,8 @@ export default function Anuncio2() {
                   variant="boton1"
                   colorScheme="brand"
                   textTransform="uppercase"
-                  w="fit-content"
+                  w={{ base: "fit-content", md: "auto" }}
+                  m={{ base: "20px", md: "5px" }}
                 >
                   Iniciar
                 </Button>
@@ -92,50 +90,8 @@ export default function Anuncio2() {
           </Flex>
         </Flex>
       </Box>
-      <Box
-        position="absolute"
-        bottom="1750px"
-        left="50%"
-        transform="translateX(-50%)"
-        textAlign="center"
-        width="100%"
-        zIndex={100}
-      >
-        <IconButton
-          aria-label="Ver más abajo"
-          icon={<FaChevronDown />}
-          size="lg"
-          colorScheme="brand"
-          variant="outline"
-          isRound
-          css={{
-            animation: "glow 1.5s ease-in-out infinite alternate",
-          }}
-          onClick={handleScroll}
-        />
-      </Box>
-      <Box
-        position="absolute"
-        bottom="980px"
-        left="49%"
-        transform="translateX(-50%)"
-        textAlign="center"
-        width="100%"
-        zIndex={100}
-      >
-        <IconButton
-          aria-label="Ver más abajo"
-          icon={<FaChevronDown />}
-          size="lg"
-          colorScheme="gray"
-          variant="outline"
-          isRound
-          css={{
-            animation: "glow 1.5s ease-in-out infinite alternate",
-          }}
-          onClick={handleScroll}
-        />
-      </Box>
+      
+    
     </chakra.header>
   );
 }
